Fix race during build of python bindings
The CMake documentation says about add_custom_command that the output must
not be used as dependency in more than one target, but this was the case
here: Both the doc_i target and the SWIG target depended on ftdi1_doc.i.
Make the SWIG target depend on doc_i instead.

This fixes that ftdi1_doc.i was built twice with parallel builds, leading
to random errors.

Signed-off-by: Fabian Vogt <fvogt@suse.de>